fix: consolidate namingschema config
